server/requirements.txt
Django==5.0.4
psycopg==3.1.8
djangorestframework==3.15.1
beautifulsoup4==4.12.3
requests==2.31.0
drf-spectacular==0.27.2